% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_genename.R
\name{add_genename}
\alias{add_genename}
\title{add_genename}
\usage{
add_genename(species, mirror = "useast", data)
}
\arguments{
\item{species}{species of the sample}

\item{mirror}{mirror for useEnsembl}

\item{data}{data object from get_deseq_res}
}
\description{
sets up gsea
}
