% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_to_gene.R
\name{collapse_to_gene}
\alias{collapse_to_gene}
\title{collapse_to_gene}
\usage{
collapse_to_gene(tpm, trans_info, selection = NULL)
}
\arguments{
\item{tpm}{output from input_kallisto}

\item{trans_info}{output from get_biomart}

\item{selection}{pattern selection for TPM calculation. defult = NULL. set it to "WT" to only use WT data for}
}
\description{
collapses transcript level info to gene level info my taking the most abundant transcript
}
