% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_plot_qc.R
\name{deseq_plot_qc}
\alias{deseq_plot_qc}
\title{deseq_plot_qc}
\usage{
deseq_plot_qc(dds)
}
\arguments{
\item{directory}{directory where raw output from feature counts are kept}
}
\description{
plots PCA
}
