% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gsea.R
\name{do_gsea}
\alias{do_gsea}
\title{do_gsea}
\usage{
do_gsea(GO_terms, data, gsea_lim = c(20, 500))
}
\arguments{
\item{GO_terms}{object from getGOterms function}

\item{data}{data from DEseq2}

\item{gsea_lim}{lower and upper limit for number of genes per GO term}
}
\description{
sets up gsea
}
