% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deseq_res.R
\name{get_deseq_res}
\alias{get_deseq_res}
\title{get_deseq_res}
\usage{
get_deseq_res(dds, shrinkage)
}
\arguments{
\item{dds}{deseq object}

\item{shrinkage}{shrinkage method}
}
\description{
exports DEseq2 results
}
