% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_featurecounts.R
\name{input_featurecounts}
\alias{input_featurecounts}
\title{input_featurecounts}
\usage{
input_featurecounts(directory)
}
\arguments{
\item{directory}{directory where raw output from feature counts are kept}
}
\description{
takes directory where feature count results are kept and loads them. add _<sample name> to the file names
}
