% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_kallisto.R
\name{input_kallisto}
\alias{input_kallisto}
\title{input_kallisto}
\usage{
input_kallisto(directory)
}
\arguments{
\item{directory}{directory where tsv output from kallisto is kept}
}
\description{
takes directory where feature count results are kept and loads them. add _<sample name> to the file names
}
