% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conversion_table.R
\name{make_conversion_table}
\alias{make_conversion_table}
\title{make_conversion_table}
\usage{
make_conversion_table(species, species_1)
}
\arguments{
\item{species}{name of biomart species mouse = mmusculus_gene_ensembl, human = hsapiens_gene_ensembl}

\item{species_1}{name of biomart species mouse = mmusculus_gene_ensembl, human = hsapiens_gene_ensembl}
}
\description{
grabs biomart info
}
