% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{networks}
\alias{networks}
\title{network}
\usage{
networks(data, full, direc, threshold)
}
\arguments{
\item{threshold}{padjust threshold for considering node}

\item{direction}{positive terms or negative terms}

\item{selection}{pattern selection for TPM calculation. defult = NULL. set it to "WT" to only use WT data for}
}
\description{
network
}
