% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_deseq.R
\name{prep_for_deseq}
\alias{prep_for_deseq}
\title{prep_for_deseq}
\usage{
prep_for_deseq(input, pattern = NULL)
}
\arguments{
\item{input}{needs to be an output from input_featurecounts}

\item{patter}{pattern to grab. defult is NULL and will return the whole input, if pattern is supplied it will return all columns with the pattern}
}
\description{
prepares count data for DESeq
}
